Re-org command definitions
